requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.2
langchain-community>=0.0.10
tavily-python>=0.2.2
wikipedia>=1.4.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
